requirement.txt
appdirs==1.4.3
asgiref==3.2.7
attrs==19.3.0
Automat==20.2.0
boto3==1.12.37
botocore==1.15.37
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.5
django-allauth==0.41.0
django-cleanup==4.0.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-swagger==2.0.7
django-storages==1.9.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
docutils==0.15.2
filelock==3.0.12
gunicorn==20.0.4
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.9.5
Markdown==3.2.1
MarkupSafe==1.1.1
mysql-connector==2.2.9
oauthlib==3.1.0
openapi-codec==1.3.2
Pillow==7.1.1
pipenv==2018.11.26
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
service-identity==18.1.0
simplejson==3.17.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
virtualenv==20.0.16
virtualenv-clone==0.5.4
whitenoise==5.0.1
zope.interface==5.1.0
